// src/app/generated/graphql.ts
export class AuthenticateUserCustomerGQL extends Apollo.Mutation<
  AuthenticateUserCustomer.Mutation,
  AuthenticateUserCustomer.Variables
> {
  document: any = gql`
    mutation authenticateUserCustomer($email: String!, $password: String!) {
      authenticateUserCustomer(input: { email: $email, password: $password }) {
        jwtToken
      }
    }
  `;
}
// ...
export class CurrentCustomerGQL extends Apollo.Query<
  CurrentCustomer.Query,
  CurrentCustomer.Variables
> {
  document: any = gql`
    query currentCustomer {
      currentCustomer {
        id
        firstName
        lastName
        email
        stripeId
      }
    }
  `;
}
// ...
export class RegisterUserCustomerGQL extends Apollo.Mutation<
  RegisterUserCustomer.Mutation,
  RegisterUserCustomer.Variables
> {
  document: any = gql`
    mutation registerUserCustomer(
      $firstName: String!
      $lastName: String!
      $email: String!
      $password: String!
    ) {
      registerUserCustomer(
        input: {
          firstName: $firstName
          lastName: $lastName
          email: $email
          password: $password
        }
      ) {
        customer {
          id
        }
      }
    }
  `;
}
// ...
export class UpdateCustomerByIdGQL extends Apollo.Mutation<
  UpdateCustomerById.Mutation,
  UpdateCustomerById.Variables
> {
  document: any = gql`
    mutation updateCustomerById(
      $customerId: UUID!
      $firstName: String!
      $lastName: String!
      $stripeId: String!
    ) {
      updateCustomerById(
        input: {
          id: $customerId
          customerPatch: {
            firstName: $firstName
            lastName: $lastName
            stripeId: $stripeId
          }
        }
      ) {
        customer {
          firstName
          lastName
          id
          stripeId
        }
      }
    }
  `;
}
// ...
export type Maybe<T> = T | null;
// ...
export type Uuid = any;
// ...
export type JwtToken = any;
// ...
  export type Variables = {
    customerId: Uuid;
  };

  export type Query = {
    __typename?: "Query";

    allAddresses: Maybe<AllAddresses>;
  };

  export type AllAddresses = {
    __typename?: "AddressesConnection";

    totalCount: Maybe<number>;

    nodes: (Maybe<Nodes>)[];
  };
// ...
export namespace AuthenticateUserCustomer {
  export type Variables = {
    email: string;
    password: string;
  };

  export type Mutation = {
    __typename?: "Mutation";

    authenticateUserCustomer: Maybe<AuthenticateUserCustomer>;
  };

  export type AuthenticateUserCustomer = {
    __typename?: "AuthenticateUserCustomerPayload";

    jwtToken: Maybe<JwtToken>;
  };
}
// ...
export namespace CurrentCustomer {
  export type Variables = {};

  export type Query = {
    __typename?: "Query";

    currentCustomer: Maybe<CurrentCustomer>;
  };

  export type CurrentCustomer = {
    __typename?: "CustomerType";

    id: Maybe<Uuid>;

    firstName: Maybe<string>;

    lastName: Maybe<string>;

    email: Maybe<string>;

    stripeId: Maybe<string>;
  };
}
// ...
export namespace RegisterUserCustomer {
  export type Variables = {
    firstName: string;
    lastName: string;
    email: string;
    password: string;
  };

  export type Mutation = {
    __typename?: "Mutation";

    registerUserCustomer: Maybe<RegisterUserCustomer>;
  };

  export type RegisterUserCustomer = {
    __typename?: "RegisterUserCustomerPayload";

    customer: Maybe<Customer>;
  };

  export type Customer = {
    __typename?: "Customer";

    id: Uuid;
  };
}
// ...
export namespace UpdateCustomerById {
  export type Variables = {
    customerId: Uuid;
    firstName: string;
    lastName: string;
    stripeId: string;
  };

  export type Mutation = {
    __typename?: "Mutation";

    updateCustomerById: Maybe<UpdateCustomerById>;
  };

  export type UpdateCustomerById = {
    __typename?: "UpdateCustomerPayload";

    customer: Maybe<Customer>;
  };

  export type Customer = {
    __typename?: "Customer";

    firstName: Maybe<string>;

    lastName: Maybe<string>;

    id: Uuid;

    stripeId: Maybe<string>;
  };
}